fixes to mr
