+++ .gitignore
@@ -51,7 +51,6 @@ coverage.xml
 *.py,cover
 .hypothesis/
 .pytest_cache/
-snapshots/
 
 # Translations
 *.mo
