+++ src/app/schemas/author.py
@@ -1,0 +1,18 @@
+from typing import Optional
+
+from pydantic import BaseModel
+
+
+class AuthorBase(BaseModel):
+    name: str
+
+
+class AuthorCreate(AuthorBase):
+    id: Optional[int]
+
+
+class Author(AuthorBase):
+    id: int
+
+    class Config:
+        orm_mode = True
